# services/payment/internal/http/oapi/api.gen.go
// Package oapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package oapi

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
	externalRef0 "github.com/teamdsb/tmo/services/payment/internal/http/oapi/common"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// AlipayPayCreateResponse defines model for AlipayPayCreateResponse.
type AlipayPayCreateResponse struct {
	OrderId   openapi_types.UUID     `json:"orderId"`
	PayParams map[string]interface{} `json:"payParams"`
	TradeNo   string                 `json:"tradeNo"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = externalRef0.ErrorResponse

// WechatPayCreateResponse defines model for WechatPayCreateResponse.
type WechatPayCreateResponse struct {
	NonceStr  string             `json:"nonceStr"`
	OrderId   openapi_types.UUID `json:"orderId"`
	PaySign   string             `json:"paySign"`
	PrepayId  string             `json:"prepayId"`
	TimeStamp string             `json:"timeStamp"`
}

// Forbidden defines model for Forbidden.
type Forbidden = ErrorResponse

// PostPaymentsAlipayCreateJSONBody defines parameters for PostPaymentsAlipayCreate.
type PostPaymentsAlipayCreateJSONBody struct {
	OrderId openapi_types.UUID `json:"orderId"`
}

// PostPaymentsAlipayCreateParams defines parameters for PostPaymentsAlipayCreate.
type PostPaymentsAlipayCreateParams struct {
	IdempotencyKey *string `json:"Idempotency-Key,omitempty"`
}

// PostPaymentsWechatCreateJSONBody defines parameters for PostPaymentsWechatCreate.
type PostPaymentsWechatCreateJSONBody struct {
	OrderId openapi_types.UUID `json:"orderId"`
}

// PostPaymentsWechatCreateParams defines parameters for PostPaymentsWechatCreate.
type PostPaymentsWechatCreateParams struct {
	IdempotencyKey *string `json:"Idempotency-Key,omitempty"`
}

// PostPaymentsWechatNotifyJSONBody defines parameters for PostPaymentsWechatNotify.
type PostPaymentsWechatNotifyJSONBody map[string]interface{}

// PostPaymentsAlipayCreateJSONRequestBody defines body for PostPaymentsAlipayCreate for application/json ContentType.
type PostPaymentsAlipayCreateJSONRequestBody PostPaymentsAlipayCreateJSONBody

// PostPaymentsWechatCreateJSONRequestBody defines body for PostPaymentsWechatCreate for application/json ContentType.
type PostPaymentsWechatCreateJSONRequestBody PostPaymentsWechatCreateJSONBody

// PostPaymentsWechatNotifyJSONRequestBody defines body for PostPaymentsWechatNotify for application/json ContentType.
type PostPaymentsWechatNotifyJSONRequestBody PostPaymentsWechatNotifyJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create Alipay payment for an order (reserved, behind feature flag)
	// (POST /payments/alipay/create)
	PostPaymentsAlipayCreate(c *gin.Context, params PostPaymentsAlipayCreateParams)
	// Create WeChat payment for an order (behind feature flag)
	// (POST /payments/wechat/create)
	PostPaymentsWechatCreate(c *gin.Context, params PostPaymentsWechatCreateParams)
	// WeChat pay callback (no auth, signature verified)
	// (POST /payments/wechat/notify)
	PostPaymentsWechatNotify(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PostPaymentsAlipayCreate operation middleware
func (siw *ServerInterfaceWrapper) PostPaymentsAlipayCreate(c *gin.Context) {

	var err error

	c.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params PostPaymentsAlipayCreateParams

	headers := c.Request.Header

	// ------------- Optional header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for Idempotency-Key, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter Idempotency-Key: %w", err), http.StatusBadRequest)
			return
		}

		params.IdempotencyKey = &IdempotencyKey

	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostPaymentsAlipayCreate(c, params)
}

// PostPaymentsWechatCreate operation middleware
func (siw *ServerInterfaceWrapper) PostPaymentsWechatCreate(c *gin.Context) {

	var err error

	c.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params PostPaymentsWechatCreateParams

	headers := c.Request.Header

	// ------------- Optional header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for Idempotency-Key, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter Idempotency-Key: %w", err), http.StatusBadRequest)
			return
		}

		params.IdempotencyKey = &IdempotencyKey

	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostPaymentsWechatCreate(c, params)
}

// PostPaymentsWechatNotify operation middleware
func (siw *ServerInterfaceWrapper) PostPaymentsWechatNotify(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostPaymentsWechatNotify(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/payments/alipay/create", wrapper.PostPaymentsAlipayCreate)
	router.POST(options.BaseURL+"/payments/wechat/create", wrapper.PostPaymentsWechatCreate)
	router.POST(options.BaseURL+"/payments/wechat/notify", wrapper.PostPaymentsWechatNotify)
}